+++ src/Application/Adapter/NotificationSubscriptionsInterface.php
@@ -5,6 +5,6 @@ namespace Proximum\Vimeet\Application\Adapter;
 
 interface NotificationSubscriptionsInterface
 {
-    public function getSubscriptions(int $eventId, int $userId): array;
+    public function getSubscriptions(int $eventId, ?int $userId, string $topicFilter = null): array;
     public function getStreamSubscriptionsCount(int $happeningId): int;
 }