+++ src/Infrastructure/Adapter/Mercure/NotificationSubscriptions.php
@@ -8,7 +8,6 @@ use Firebase\JWT\JWT;
 use Proximum\Vimeet\Application\Adapter\HttpAdapterInterface;
 use Proximum\Vimeet\Application\Adapter\NotificationSubscriptionsInterface;
 
-
 class NotificationSubscriptions extends AbstractNotification implements NotificationSubscriptionsInterface
 {
     /** @var string */
@@ -30,13 +29,16 @@ class NotificationSubscriptions extends AbstractNotification implements Notifica
         $this->httpAdapter = $httpAdapter;
     }
 
-    public function getSubscriptions(int $eventId, int $userId): array
+    public function getSubscriptions(int $eventId, ?int $userId, string $topicFilter = null): array
     {
         $result = $this->getSubscriptionsRaw();
 
         $users = [];
 
         foreach ($result['subscriptions'] as $subscription) {
+            if ($topicFilter && $subscription['topic'] !== $topicFilter) {
+                continue;
+            }
             if (!isset($subscription['payload']['eventId']) || $subscription['payload']['eventId'] != $eventId) {
                 continue;
             }
@@ -44,7 +46,7 @@ class NotificationSubscriptions extends AbstractNotification implements Notifica
                 continue;
             }
             // exclude current user
-            if ($userId === $subscription['payload']['userId']) {
+            if ($userId && $userId === $subscription['payload']['userId']) {
                 continue;
             }
             $users[$subscription['payload']['userId']] = $subscription['payload'];