+++ src/Ui/Bundle/EventBundle/Resources/assets/js/main.js
@@ -38,6 +38,7 @@ import { showSpinnerOnSubmit } from './components/form/_ShowSpinnerOnSubmit';
 import SheetVideo from './components/_SheetVideo';
 import UserConnectionRegister from './components/_UserConnectionRegister';
 import NotificationToast from './components/_NotificationToast';
+import NotificationCallVisio from './components/_NotificationCallVisio';
 import initNetworking from './networking'
 
 import 'bootstrap';
@@ -288,7 +289,9 @@ function init(target) {
     const userConnectionRegister = new UserConnectionRegister();
     userConnectionRegister.connect();
 
-    initNetworking(document, userConnectionRegister);
+    const notificationCallVisio = new NotificationCallVisio(target.querySelector('[data-notification-call-visio]'), userConnectionRegister);
+
+    initNetworking(document, userConnectionRegister, notificationCallVisio);
 
     new NotificationToast(target.querySelector('[data-notification-toast]'), userConnectionRegister);