+++ CHANGELOG.md
@@ -4,6 +4,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [2.66.1] - 2020-11-05
+### Fixed
+- MV264 - Networking tab only validated or accepted
+
+## [2.66.0] - 2020-11-04
+### Added
+- VIMEET-2223 - Add call visio on private chat (tab networking).
+- VIMEET-2229 - Add field date call visio in Admin.
+- VIMEET-2260 - Add button custom 2.
 
 ## [2.65.2] - 2020-11-02
 ### Fixed
@@ -16,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [2.65.0] - 2020-10-29
 ### Added
 - VIMEET-2225 - Add related product to video template objects
+>>>>>>> master
 
 ## [2.64.0] - 2020-10-28
 ### Fixed
@@ -42,6 +52,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - VIMEET-2219 - Apply filter on users added on the fly
 - Add sheet card tag for datetime
 - Fix impersonation route for role with ROLE_ALLOWED_TO_SWITCH
+- VIMEET-2229 - Add call visio networking tab by key date
 
 ## [2.61.0] - 2020-10-15
 ### Added